feat: add microfips-l2cap-test crate, fix clippy warnings, update AGENTS.md with FIPS ble-transport-reliability audit
- Add microfips-l2cap-test crate with 50 BLE L2CAP protocol compatibility tests
- Add l2cap-test to CI (test job, fmt, clippy)
- Increase L2CAP_FRAME_CAP 512→768 (max that fits ESP32 DRAM budget)
- Increase supervision_timeout 400ms→4000ms (prevent BLE link disconnects)
- Fix clippy warnings: checked_div, unnecessary_cast, missing Default impls
- Update AGENTS.md: ble-transport-reliability branch audit, pubkey exchange format,
  FIPS issue cross-references, upstream compatibility notes
- Keep L2CAP MTU at 2048 (matches FIPS default, FRAME_CAP handles buffering)

+++ crates/microfips-protocol/src/peer_policy.rs
@@ -48,7 +48,7 @@ impl PeerPolicy {
             return PolicyVerdict::Allow;
         };
 
-        let elapsed_ms = now.as_millis().saturating_sub(last_connect.as_millis()) as u64;
+        let elapsed_ms = now.as_millis().saturating_sub(last_connect.as_millis());
 
         let failure_backoff_ms = if self.consecutive_failures >= MAX_CONSECUTIVE_FAILURES {
             MAX_RECONNECT_MS
@@ -73,7 +73,7 @@ impl PeerPolicy {
     pub fn check_frame_rate(&mut self, now: Instant) -> PolicyVerdict {
         let elapsed_ms = now
             .as_millis()
-            .saturating_sub(self.frame_window_start.as_millis()) as u64;
+            .saturating_sub(self.frame_window_start.as_millis());
 
         if elapsed_ms >= FRAME_RATE_WINDOW_MS {
             self.frame_window_start = now;
